Reorganize imports to make Python 3 support easier.
